feat(handle): add Widget series types
